fix: fix bugs in train.py and augment.py as well as refactor the lambda function in train.py

+++ src/utils/augment.py
@@ -95,7 +95,8 @@ class RandomFreqPerturbation:
         radius_map = torch.sqrt(grid_x**2 + grid_y**2)
 
         scale = random.uniform(*self.scale_range)
-        mask = torch.ones((h, w), device=device, dtype=dtype)
+        # 使用 freq 的 dtype (complex) 而非 tensor 的 dtype (float)
+        mask = torch.ones((h, w), device=device, dtype=freq.dtype)
         mask = mask + (radius_map >= self.radius) * (scale - 1.0)
         mask = mask.unsqueeze(0)
